feat(client): use groups as sub-object
[Breaking change]: groups exposed by service are now used as sub-object of mini-service.

Also refactor tests structure for smaller files
