+++ docs/client.js.html
@@ -22,7 +22,7 @@
 <label for="nav-trigger" class="overlay"></label>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Classes</h3><ul><li><a href="Client.html">Client</a><ul class='methods'><li data-type='method'><a href="Client.html#init">init</a></li></ul></li></ul><h3>Modules</h3><ul><li><a href="module-mini-client.html">mini-client</a><ul class='methods'><li data-type='method'><a href="module-mini-client.html#~createClient">createClient</a></li></ul></li><li><a href="module-register.html">register</a><ul class='methods'><li data-type='method'><a href="module-register.html#.registerFromServer">registerFromServer</a></li><li data-type='method'><a href="module-register.html#.registerLocal">registerLocal</a></li></ul></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Classes</h3><ul><li><a href="Client.html">Client</a><ul class='methods'><li data-type='method'><a href="Client.html#init">init</a></li></ul></li></ul><h3>Modules</h3><ul><li><a href="module-mini-client.html">mini-client</a><ul class='methods'><li data-type='method'><a href="module-mini-client.html#~createClient">createClient</a></li></ul></li><li><a href="module-register.html">register</a><ul class='methods'><li data-type='method'><a href="module-register.html#.registerFromServer">registerFromServer</a></li><li data-type='method'><a href="module-register.html#.registerLocal">registerLocal</a></li><li data-type='method'><a href="module-register.html#~computeGroup">computeGroup</a></li></ul></li></ul>
 </nav>
 
 <div id="main">
@@ -109,20 +109,47 @@ class Client {
           return undefined
         }
         // creates a function that will get exposed API from remote server
-        return (...args) => {
-          logger.debug(`during ${propKey}, connect to ${remote} to get exposed apis`)
-          return registerFromServer(target, remote, logger)
-            .then(() => {
-              // delete trap to boost performance: next calls will directly search into 'this'
-              delete traps.get
-              logger.debug('remote client ready')
-              // now, invoke the API with initial arguments, but only if it exists
-              if (!(propKey in target)) {
-                throw new TypeError(`${propKey} is not a function`)
-              }
-              return target[propKey](...args)
-            })
-        }
+        return new Proxy(() => { /* esling no-empty: 0 */ }, {
+          get(_1, subPropKey) {
+            // reserved keyword that aren't defined must be undefined (thenable false-positive)
+            if (reserved.includes(subPropKey)) {
+              /* eslint no-undefined: 0 */
+              return undefined
+            }
+            return (...args) => {
+              logger.debug(`during ${propKey}.${subPropKey}, connect to ${remote} to get exposed apis`)
+              return registerFromServer(target, remote, logger)
+                .then(() => {
+                  // delete trap to boost performance: next calls will directly search into 'this'
+                  delete traps.get
+                  logger.debug('remote client ready')
+                  // now, invoke the API with initial arguments, but only group exists
+                  if (!(propKey in target)) {
+                    throw new TypeError(`Cannot read property '${subPropKey}' of undefined`)
+                  }
+                  // and if property is defined in group
+                  if (!(subPropKey in target[propKey])) {
+                    throw new TypeError(`${propKey}.${subPropKey} is not a function`)
+                  }
+                  return target[propKey][subPropKey](...args)
+                })
+            }
+          },
+          apply(_1, _2, args) {
+            logger.debug(`during ${propKey}, connect to ${remote} to get exposed apis`)
+            return registerFromServer(target, remote, logger)
+              .then(() => {
+                // delete trap to boost performance: next calls will directly search into 'this'
+                delete traps.get
+                logger.debug('remote client ready')
+                // now, invoke the API with initial arguments, but only if defined in group
+                if (!(propKey in target)) {
+                  throw new TypeError(`${propKey} is not a function`)
+                }
+                return target[propKey](...args)
+              })
+          }
+        })
       }
     }
     return remote ? new Proxy(this, traps) : this
