+++ lib/register.js
@@ -16,6 +16,32 @@ const {
  */
 
 /**
+ * Created operation should be regrouped according to their own group
+ *
+ * Registering operations could occur:
+ * - on root context when operation belongs to group named afer the service
+ * - on group subobject when operation belongs to a different group
+ *
+ * @param {Object} context root context in which operations will be created
+ * @param {String} group group name currently processed
+ * @param {Object} opts options describing local/remote service
+ * @param {String} opts.name of current service
+ * @returns {Object} in which operations could be registered as functions
+ */
+const computeGroup = (context, group, opts) => {
+  // use group if it's not the service name
+  if (group !== opts.name) {
+    // creates group if not present yet
+    if (!(group in context)) {
+      context[group] = {}
+    }
+    return context[group]
+  }
+  return context
+}
+
+
+/**
  * @summary Ask a given server for APIs to register into the given context (remote client)
  *
  * @param {Object} client   in which exposed api will be registered
@@ -35,7 +61,7 @@ exports.registerFromServer = (client, url, logger) =>
     // add one method to client per exposed api
     exposed.apis.forEach(({group, path, id, params}) => {
       const method = params.length ? 'POST' : 'GET'
-      client[id] = (...args) =>
+      computeGroup(client, group, exposed)[id] = (...args) =>
         request({
           method,
           uri: `${url}${path}`,
@@ -87,19 +113,19 @@ exports.registerLocal = (context, opts, logger) => {
   try {
     const {groups, groupOpts} = extractGroups(opts)
     return [Promise.resolve()].concat(groups)
-      .reduce((previous, {name, init}) =>
+      .reduce((previous, {name: group, init}) =>
         previous
           .then(() => {
-            const initialized = init(Object.assign({logger}, groupOpts[name]))
+            const initialized = init(Object.assign({logger}, groupOpts[group]))
             if (Promise.resolve(initialized) !== initialized) {
-              throw new Error(`${name} init() method didn't returned a promise`)
+              throw new Error(`${group} init() method didn't returned a promise`)
             }
             return initialized
           })
           .then(apis => {
             if (!isApi(apis)) return
             for (const id in apis) {
-              const validate = extractValidate(id, apis, groupOpts[name])
+              const validate = extractValidate(id, apis, groupOpts[group])
               // extrat param names for validation
               const params = getParamNames(apis[id])
               let schema = null
@@ -110,7 +136,7 @@ exports.registerLocal = (context, opts, logger) => {
               }
 
               // enrich context with a dedicated function
-              context[id] = (...args) => {
+              computeGroup(context, group, opts)[id] = (...args) => {
                 // adds input validation
                 if (schema) {
                   const error = validateParams(arrayToObj(args, params), schema, id, params.length)
@@ -128,8 +154,8 @@ exports.registerLocal = (context, opts, logger) => {
                   return Promise.reject(exc)
                 }
               }
+              logger.debug(`API ${id} from ${group} loaded`)
             }
-            logger.debug({apis: Object.keys(apis)}, `APIs from ${name} loaded`)
           })
       )
   } catch (err) {
